+++ scripts/bench-create.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+
+cd "../cli"
+
+# Create and publish a demo federated graph based on the subgraphs in the demo folder
+
+export COSMO_API_KEY=cosmo_669b576aaadc10ee1ae81d9193425705
+export COSMO_API_URL=http://localhost:3001
+
+pnpm wgc federated-graph create bench --label-matcher team=Bench --routing-url http://localhost:3003/graphql
+
+pnpm wgc subgraph create bench-accounts --label team=Bench --routing-url http://localhost:4001/graphql
+pnpm wgc subgraph create bench-reviews --label team=Bench --routing-url http://localhost:4002/graphql
+pnpm wgc subgraph create bench-products --label team=Bench --routing-url http://localhost:4003/graphql
+pnpm wgc subgraph create bench-inventory --label team=Bench --routing-url http://localhost:4004/graphql
+
+cd "../scripts"
+
+./bench-update.sh
+
+cd "../cli"
+
+pnpm wgc federated-graph create-token bench --name mytoken